pyproject.toml
[project]
name = "news-aggregator-agent"
version = "1.0.0"
description = "AI-powered news aggregation agent with multi-source integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-experimental>=0.0.50",
    "fastapi>=0.104.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "langchain_tavily>=0.0.1",
    "langchain_groq>=0.0.1",
    "langchain_openai>=0.0.5",
    "langgraph>=0.0.20",
    "python-dateutil>=2.8.0",
    "colorama>=0.4.0",
    "rich>=13.0.0",
]
